LayoutTests/webaudio/panner-set-model-expected.txt

Test if panningModel can be set and read.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Panner set to EQUALPOWER model and read correctly.
PASS Panner set to HRTF model and read correctly.
PASS Setting panner model to SOUNDFIELD correctly throws exception because it is not implemented.
PASS Illegal panner model correctly throws exception.
PASS Panning model tests passed.
PASS successfullyParsed is true

TEST COMPLETE
